Generate a random poster when button is clicked
Instead of declaring a new var we reassigned the current poster
var. Moved randomPoster var inside renderRandomPoster, and changed
it to currentPoster to redeclare. There was 1 instance being made, and
we needed to make new instances of the var. The function was not working
with the var outside of the renderRandomPoster.